rearrange lib structure
