--- init.rb
@@ -1,1 +1,0 @@
-ActionView::Base.send :include, MicroformatsHelper
